+++ app/routes/components/Input/demos/Demo02.tsx
@@ -1,5 +1,5 @@
 import { Input } from "@/components/Input";
 
 export default function Demo() {
-  return <Input placeholder="Type something..." />;
+  return <Input variant="filled" placeholder="Search..." />;
 }