+++ app/routes/components/Input/demos/Demo04.tsx
@@ -1,5 +1,5 @@
 import { Input } from "@/components/Input";
 
 export default function Demo() {
-  return <Input placeholder="Type something..." />;
+  return <Input disabled placeholder="Disabled input" />;
 }