+++ app/routes/components/Input/demos/Demo05.tsx
@@ -1,0 +1,12 @@
+import { Flex } from "@/components/Flex";
+import { Input } from "@/components/Input";
+
+export default function Demo() {
+  return (
+    <Flex dir="column" gap="1rem">
+      <Input size="sm" placeholder="Type something..." />
+      <Input size="md" placeholder="Type something..." />
+      <Input size="lg" placeholder="Type something..." />
+    </Flex>
+  );
+}