+++ src/components/Input/CHANGELOG.md
@@ -10,3 +10,4 @@
   component to add elements inside the input field on the left or right side.
   Adjust padding based on presence of these sections.
 - 2025-10-20: Support `disabled` prop.
+- 2025-10-26: Support `variant` prop. Update styles to make it look better.